requirements.txt
colorama==0.4.6
config==0.5.1
loguru==0.7.2
numpy==1.26.4
pandas==2.2.1
pathlib==1.0.1
plum-py==0.8.7
PyExifTool==0.4.13
python-dateutil==2.8.2
python-slugify==8.0.4
python-xmp-toolkit==2.0.2
pytz==2024.1
pyxmp==0.1.1
PyYAML==6.0.1
six==1.16.0
text-unidecode==1.3
tqdm==4.66.2
tzdata==2024.1
